ai_engine/requirements.txt
ultralytics
paho-mqtt
numpy
opencv-python-headless
